Compiler warnings on Linux

+++ src/radar_pi.cpp
@@ -1244,7 +1244,7 @@ bool radar_pi::LoadConfig(void) {
     M_SETTINGS.radar_count = v;
 
     size_t n = 0;
-    for (size_t r = 0; r < M_SETTINGS.radar_count; r++) {
+    for (int r = 0; r < M_SETTINGS.radar_count; r++) {
       RadarInfo *ri = m_radar[n];
       pConf->Read(wxString::Format(wxT("Radar%dType"), r), &s, "unknown");
       ri->m_radar_type = RT_MAX;  // = not used
@@ -1425,7 +1425,7 @@ bool radar_pi::SaveConfig(void) {
     pConf->Write(wxT("ColourPPIBackground"), m_settings.ppi_background_colour.GetAsString());
     pConf->Write(wxT("RadarCount"), m_settings.radar_count);
 
-    for (size_t r = 0; r < m_settings.radar_count; r++) {
+    for (int r = 0; r < m_settings.radar_count; r++) {
       pConf->Write(wxString::Format(wxT("Radar%dType"), r), RadarTypeName[m_radar[r]->m_radar_type]);
 
       wxString addr;
